Make chart without margins. Disable buttons for some chartTypes Refactor and delete old classes
