--- UnstoppableWallet/UnstoppableWallet/Modules/Chart/CoinInfoMap.swift
@@ -1,285 +1,0 @@
-import Foundation
-
-class CoinInfoMap {
-
-    static let data: [String: CoinInfo] = [
-
-
-        "$BASED":     CoinInfo(supply: 100_000,         startDate: nil,          website: "https://based.money"),
-        "0chain":     CoinInfo(supply: nil,             startDate: nil,          website: "https://0chain.net"),
-        "1INCH":      CoinInfo(supply: 1_500_000_000,   startDate: nil,          website: "https://1inch.exchange"),
-        "AAVE":       CoinInfo(supply: 16_000_000,      startDate: "02/10/2020", website: "https://aave.com"),
-        "AAVEDAI":    CoinInfo(supply: nil,             startDate: "08/01/2020", website: "https://aave.com"),
-        "AKRO":       CoinInfo(supply: 4_000_000_000,   startDate: "06/07/2019", website: "https://akropolis.io"),
-        "ALBT":       CoinInfo(supply: 1_000_000_000,   startDate: nil,          website: "https://www.allianceblock.io"),
-        "ALPHA":      CoinInfo(supply: 1_000_000,       startDate: nil,          website: "https://alphafinance.io"),
-        "AMON":       CoinInfo(supply: 1_666_666_667,   startDate: "13/04/2018", website: "https://amon.tech/"),
-        "AMPL":       CoinInfo(supply: 780_675_794,     startDate: "14/06/2019", website: "https://www.ampleforth.org"),
-        "ANJ":        CoinInfo(supply: 128_645_461,     startDate: "07/01/2020", website: "https://anj.aragon.org"),
-        "ANKR":       CoinInfo(supply: 10_000_000_000,  startDate: "21/02/2019", website: "https://www.ankr.com"),
-        "ANT":        CoinInfo(supply: 39_609_524,      startDate: "05/05/2017", website: "https://aragon.one"),
-        "API3":       CoinInfo(supply: nil,             startDate: nil,          website: "https://api3.org"),
-        "APY":        CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://apy.finance"),
-        "AST":        CoinInfo(supply: 500_000_000,     startDate: "07/01/2020", website: "https://www.airswap.io"),
-        "AUC":        CoinInfo(supply: nil,             startDate: nil,          website: "https://auctus.org"),
-        "BAC":        CoinInfo(supply: 50_001,          startDate: nil,          website: "https://basis.cash"),
-        "BADGER":     CoinInfo(supply: 21_000_000,      startDate: nil,          website: "https://badger.finance"),
-        "BAL":        CoinInfo(supply: 35_725_000,      startDate: "20/06/2020", website: "https://balancer.finance"),
-        "BAND":       CoinInfo(supply: 100_000_000,     startDate: "09/09/2019", website: "https://bandprotocol.com"),
-        "BASE":       CoinInfo(supply: 65_370_936,      startDate: nil,          website: "https://www.baseprotocol.org"),
-        "BAT":        CoinInfo(supply: 1_500_000_000,   startDate: "31/05/2017", website: "https://basicattentiontoken.org"),
-        "BCH":        CoinInfo(supply: 21_000_000,      startDate: "01/08/2017", website: "https://www.bitcoincash.org"),
-        "BID":        CoinInfo(supply: 4_000_000_000,   startDate: nil,          website: "https://bidaochain.com"),
-        "BLT":        CoinInfo(supply: 150_000_000,     startDate: "10/11/2017", website: "https://bloom.co"),
-        "BNB":        CoinInfo(supply: 187_536_713,     startDate: "27/06/2017", website: "https://www.binance.com"),
-        "BNT":        CoinInfo(supply: nil,             startDate: "13/02/2017", website: "https://bancor.network"),
-        "BOND":       CoinInfo(supply: nil,             startDate: nil,          website: "https://barnbridge.com"),
-        "BONDLY":     CoinInfo(supply: 100_000_000_000, startDate: nil,          website: "https://www.bondly.finance"),
-        "BTC":        CoinInfo(supply: 21_000_000,      startDate: "03/01/2009", website: "https://bitcoin.org/en"),
-        "BTCB":       CoinInfo(supply: 9_001,           startDate: "17/06/2019", website: "https://bitcoin.org/en/"),
-        "BUSD":       CoinInfo(supply: 28_603_822,      startDate: "10/09/2019", website: "https://www.paxos.com/busd"),
-        "BZRX":       CoinInfo(supply: 1_030_000_000,   startDate: "04/02/2019", website: "https://bzx.network"),
-        "CAS":        CoinInfo(supply: 1_000_000_000,   startDate: "12/10/2017", website: "https://cashaa.com"),
-        "CC10":       CoinInfo(supply: nil,             startDate: nil,          website: "https://indexed.finance"),
-        "CDAI":       CoinInfo(supply: 2_025_596_474,   startDate: "07/05/2019", website: "https://compound.finance"),
-        "CEL":        CoinInfo(supply: 695_658_161,     startDate: "11/04/2018", website: "https://celsius.network"),
-        "CELR":       CoinInfo(supply: 10_000_000_000,  startDate: nil,          website: "https://www.celer.network"),
-        "CHAI":       CoinInfo(supply: nil,             startDate: nil,          website: nil),
-        "CHAIN":      CoinInfo(supply: 500_000_000,     startDate: nil,          website: "https://chaingames.io"),
-        "CHSB":       CoinInfo(supply: 1_000_000_000,   startDate: "08/09/2017", website: "https://swissborg.com"),
-        "CHZ":        CoinInfo(supply: 8_888_888_888,   startDate: nil,          website: "https://www.chiliz.com"),
-        "COMBO":      CoinInfo(supply: nil,             startDate: nil,          website: "https://furucombo.app"),
-        "COMP":       CoinInfo(supply: 10_000_000,      startDate: "04/03/2020", website: "https://compound.finance"),
-        "CORE":       CoinInfo(supply: nil,             startDate: nil,          website: "https://cvault.finance"),
-        "COS":        CoinInfo(supply: 9_400_000_000,   startDate: nil,          website: "https://www.contentos.io"),
-        "COVER":      CoinInfo(supply: nil,             startDate: nil,          website: "https://www.coverprotocol.com"),
-        "CREAM":      CoinInfo(supply: 2_924_547,       startDate: nil,          website: "https://cream.finance"),
-        "CRO":        CoinInfo(supply: 100_000_000_000, startDate: "14/11/2019", website: "https://www.crypto.com/en/chain"),
-        "CRPT":       CoinInfo(supply: 99_785_291,      startDate: "28/09/2017", website: "https://crypterium.io"),
-        "CRV":        CoinInfo(supply: 1_356_236_799,   startDate: "13/08/2020", website: "https://www.curve.fi/"),
-        "CSAI":       CoinInfo(supply: 17_342_084,      startDate: "07/05/2019", website: "https://compound.finance/"),
-        "CUSDC":      CoinInfo(supply: 9_803_299_805,   startDate: "07/05/2019", website: "https://compound.finance"),
-        "CVC":        CoinInfo(supply: 1_000_000_000,   startDate: "21/06/2017", website: "https://www.civic.com"),
-        "DAI":        CoinInfo(supply: 127_759_558,     startDate: "18/11/2019", website: "https://makerdao.com/en/"),
-        "DASH":       CoinInfo(supply: 18_900_000,      startDate: "18/01/2014", website: "https://dash.org"),
-        "DDIM":       CoinInfo(supply: 1_500_000,       startDate: nil,          website: "https://duckdao.io"),
-        "DEFI5":      CoinInfo(supply: nil,             startDate: nil,          website: "https://indexed.finance"),
-        "DEFO":       CoinInfo(supply: nil,             startDate: nil,          website: "https://defhold.com"),
-        "DEGO":       CoinInfo(supply: nil,             startDate: nil,          website: "https://dego.finance"),
-        "DEUS":       CoinInfo(supply: 1_793_062,       startDate: nil,          website: "https://deus.finance/home"),
-        "DGD":        CoinInfo(supply: 2_000_000,       startDate: "28/04/2016", website: "https://www.dgx.io"),
-        "DGX":        CoinInfo(supply: 120_600,         startDate: "29/03/2018", website: "https://digix.global/dgx#/"),
-        "DHT":        CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://www.dhedge.org/"),
-        "DIA":        CoinInfo(supply: 200_000_000,     startDate: "27/11/2019", website: "https://diadata.org/"),
-        "DIP":        CoinInfo(supply: 1_000_000_000,   startDate: "31/05/2018", website: "https://www.etherisc.com/"),
-        "DMG":        CoinInfo(supply: 249_654_286,     startDate: nil,          website: "https://defimoneymarket.com"),
-        "DNT":        CoinInfo(supply: 1_000_000_000,   startDate: "08/08/2017", website: "https://district0x.io"),
-        "DOS":        CoinInfo(supply: 949_292_114,     startDate: "11/03/2019", website: "https://dos.network/"),
-        "DPI":        CoinInfo(supply: nil,             startDate: nil,          website: "https://www.indexcoop.com/dpi"),
-        "DSD":        CoinInfo(supply: 35_331_663,      startDate: nil,          website: "https://dsd.finance/app/#/"),
-        "DUSD":       CoinInfo(supply: 16_697_847,      startDate: nil,          website: "https://app.dusd.finance"),
-        "DYP":        CoinInfo(supply: nil,             startDate: nil,          website: "https://dyp.finance"),
-        "ELF":        CoinInfo(supply: 880_000_000,     startDate: "18/12/2017", website: "https://aelf.io"),
-        "ENJ":        CoinInfo(supply: 1_000_000_000,   startDate: "24/07/2017", website: "https://enjincoin.io"),
-        "ESD":        CoinInfo(supply: 443_687_674,     startDate: nil,          website: "https://emptyset.finance"),
-        "ETH":        CoinInfo(supply: nil,             startDate: "30/07/2015", website: "https://www.ethereum.org"),
-        "ETH-BEP2":   CoinInfo(supply: 215_000,         startDate: nil,          website: nil),
-        "ETHYS":      CoinInfo(supply: 500_000,         startDate: nil,          website: "https://ethereumstake.farm"),
-        "EURS":       CoinInfo(supply: nil,             startDate: "22/06/2018", website: "https://stasis.net"),
-        "FARM":       CoinInfo(supply: 471_639,         startDate: nil,          website: "https://harvest.finance"),
-        "FLASH":      CoinInfo(supply: 900_000_000,     startDate: nil,          website: "https://flashstake.io"),
-        "FLUX":       CoinInfo(supply: 259_186,         startDate: nil,          website: "https://datamine.network"),
-        "FNK":        CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://fnktoken.com"),
-        "FOAM":       CoinInfo(supply: 1_000_000_000,   startDate: "10/10/2017", website: "https://foam.space"),
-        "FOR":        CoinInfo(supply: 1_000_000_000,   startDate: nil,          website: "https://www.theforceprotocol.com"),
-        "FRAX":       CoinInfo(supply: 130_600_430,     startDate: nil,          website: "https://frax.finance"),
-        "FSW":        CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://falconswap.com"),
-        "FUN":        CoinInfo(supply: 10_999_873_621,  startDate: "23/06/2017", website: "https://www.funfair.io"),
-        "FYZ":        CoinInfo(supply: 50_000_000,      startDate: nil,          website: "https://www.fyooz.io"),
-        "GEN":        CoinInfo(supply: 60_000_000,      startDate: "01/05/2018", website: "https://daostack.io"),
-        "GLM":        CoinInfo(supply: nil,             startDate: "17/11/2016", website: "https://golem.network"),
-        "GNO":        CoinInfo(supply: 10_000_000,      startDate: "25/04/2017", website: "https://gnosis.io/"),
-        "GRID":       CoinInfo(supply: 300_000_000 ,    startDate: "12/07/2017", website: "https://gridplus.io"),
-        "GRT":        CoinInfo(supply: 10_000_000_000,  startDate: nil,          website: "https://thegraph.com"),
-        "GST2":       CoinInfo(supply: nil,             startDate: nil,          website: "https://gastoken.io"),
-        "GTO":        CoinInfo(supply: 1_000_000_000,   startDate: "14/12/2017", website: "https://gifto.io"),
-        "GUSD":       CoinInfo(supply: nil,             startDate: "09/09/2018", website: "https://gemini.com/dollar"),
-        "HEDG":       CoinInfo(supply: 1_000_000_000,   startDate: "07/12/2018", website: "https://hedgetrade.com"),
-        "HEGIC":      CoinInfo(supply: nil,             startDate: nil,          website: "https://www.hegic.co"),
-        "HEZ":        CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://hermez.io"),
-        "HH":         CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://holyheld.com"),
-        "HLAND":      CoinInfo(supply: nil,             startDate: nil,          website: "https://htw.land"),
-        "HOT":        CoinInfo(supply: 177_619_433_541, startDate: "16/01/2018", website: "https://thehydrofoundation.com"),
-        "HT":         CoinInfo(supply: 500_000_000,     startDate: "22/01/2018", website: "https://www.huobi.pro"),
-        "HUSD":       CoinInfo(supply: nil,             startDate: "29/11/2018", website: "https://www.stcoins.com"),
-        "IDEX":       CoinInfo(supply: 1_000_000_000,   startDate: "28/09/2017", website: "https://auroradao.com"),
-        "IDLE":       CoinInfo(supply: 13_000_000,      startDate: nil,          website: "https://idle.finance"),
-        "IOTX":       CoinInfo(supply: 10_000_000_000,  startDate: "21/05/2018", website: "https://iotex.io/"),
-        "IRIS":       CoinInfo(supply: 2_000_000_000,   startDate: nil,          website: "https://www.irisnet.org"),
-        "KCS":        CoinInfo(supply: 176_863_551,     startDate: "15/09/2017", website: "https://www.kucoin.com/#"),
-        "KEEP":       CoinInfo(supply: 1_000_000_000,   startDate: nil,          website: "https://keep.network"),
-        "KEY":        CoinInfo(supply: 5_999_999_954 ,  startDate: "11/01/2018", website: "https://selfkey.org/"),
-        "KNC":        CoinInfo(supply: 210_590_429,     startDate: "20/09/2017", website: "https://kyber.network"),
-        "KP3R":       CoinInfo(supply: nil,             startDate: nil,          website: "https://keep3r.network"),
-        "LDO":        CoinInfo(supply: 1_000_000_000,   startDate: nil,          website: "https://stake.lido.fi/"),
-        "LEND":       CoinInfo(supply: 1_299_999_942,   startDate: "15/08/2017", website: "https://aave.com"),
-        "LGCY":       CoinInfo(supply: nil,             startDate: nil,          website: "https://lgcy.network"),
-        "LINA":       CoinInfo(supply: 10_000_000_199,  startDate: "16/09/2020", website: "https://linear.finance"),
-        "LINK":       CoinInfo(supply: 1_000_000_000,   startDate: "19/09/2017", website: "https://link.smartcontract.com"),
-        "LON":        CoinInfo(supply:  200_000_000,    startDate: nil,          website: "https://tokenlon.im/lon"),
-        "LOOM":       CoinInfo(supply: 1_000_000_000,   startDate: "03/03/2018", website: "https://loomx.io"),
-        "LPT":        CoinInfo(supply: 22_537_813,      startDate: "08/11/2018", website: "https://livepeer.org"),
-        "LQD":        CoinInfo(supply: 100_000_000,     startDate: "01/11/2015", website: "https://liquidity.network"),
-        "LRC":        CoinInfo(supply: 1_374_513_897,   startDate: "06/08/2017", website: "https://loopring.org/#/"),
-        "LTC":        CoinInfo(supply: 84_000_000,      startDate: "13/10/2011", website: "https://litecoin.com"),
-        "LTO":        CoinInfo(supply: 500_000_000,     startDate: nil,          website: "https://lto.network"),
-        "MANA":       CoinInfo(supply: 2_644_403_343,   startDate: "08/08/2017", website: "https://decentraland.org"),
-        "MATIC":      CoinInfo(supply: 10_000_000_000,  startDate: "26/03/2018", website: "https://matic.network"),
-        "MATIC-BEP2": CoinInfo(supply: 10_000_000_000,  startDate: nil,          website: nil),
-        "MCB":        CoinInfo(supply: 150_000_000,     startDate: nil,          website: "https://mcdex.io"),
-        "MCO":        CoinInfo(supply: 31_587_682,      startDate: "18/05/2017", website: "https://crypto.com/en/index.html"),
-        "MEME":       CoinInfo(supply: 28_000,          startDate: nil,          website: "https://dontbuymeme.com"),
-        "MET":        CoinInfo(supply: 12_488_971,      startDate: "23/05/2018", website: "https://www.metronome.io"),
-        "MFT":        CoinInfo(supply: 10_000_000_000,  startDate: nil,          website: "https://mainframe.com"),
-        "MIR":        CoinInfo(supply: nil,             startDate: nil,          website: "https://eth.mirror.finance"),
-        "MKR":        CoinInfo(supply: 1_000_000,       startDate: "15/08/2015", website: nil),
-        "MLN":        CoinInfo(supply: 1_524_032,       startDate: "26/01/2018", website: "https://melonprotocol.com"),
-        "MPH":        CoinInfo(supply: 304_412,         startDate: nil,          website: "https://88mph.app"),
-        "MTA":        CoinInfo(supply: 100_000_000,     startDate: "13/07/2020", website: "https://mstable.org"),
-        "MTXLT":      CoinInfo(supply: 90_000,          startDate: nil,          website: "https://tixl.me"),
-        "MUSD":       CoinInfo(supply: 100_000_000,     startDate: "05/01/2018", website: "https://mstable.org"),
-        "NEST":       CoinInfo(supply: nil,             startDate: nil,          website: "https://nestprotocol.org"),
-        "NEXO":       CoinInfo(supply: 1_000_000_000,   startDate: "29/04/2018", website: "https://nexo.io"),
-        "NMR":        CoinInfo(supply: 11_000_000,      startDate: "21/06/2017", website: "https://erasure.world"),
-        "NPXS":       CoinInfo(supply: 259_810_708_833, startDate: "27/09/2017", website: "https://pundix.com"),
-        "NXM":        CoinInfo(supply: 6_696_774,       startDate: nil,          website: "https://nexusmutual.io"),
-        "Nsure":      CoinInfo(supply: nil,             startDate: nil,          website: "https://nsure.network"),
-        "OCEAN":      CoinInfo(supply: 613_099_141,     startDate: "23/04/2019", website: "https://oceanprotocol.com"),
-        "OCTO":       CoinInfo(supply: 800_000,         startDate: nil,          website: "https://octo.fi"),
-        "OGN":        CoinInfo(supply: 1_000_000_000,   startDate: nil,          website: "https://www.originprotocol.com"),
-        "OMG":        CoinInfo(supply:  nil,            startDate: "23/06/2017", website: "https://omg.network"),
-        "ONE":        CoinInfo(supply: 12_600_000_000,  startDate: nil,          website: "https://harmony.one"),
-        "ORAI":       CoinInfo(supply: 23_037_721,      startDate: nil,          website: "https://orai.io"),
-        "ORBS":       CoinInfo(supply: 10_000_000_000,  startDate: "14/03/2018", website: "https://www.orbs.com"),
-        "ORN":        CoinInfo(supply: 100_000_000,     startDate: "13/07/2020", website: "https://orionprotocol.io"),
-        "OXT":        CoinInfo(supply: 10_000_000_000,  startDate: "03/12/2019", website: "https://www.orchid.com"),
-        "PAN":        CoinInfo(supply: nil,             startDate: nil,          website: "https://panvala.com"),
-        "PAR":        CoinInfo(supply: 999_628_334,     startDate: nil,          website: "https://www.parachutetoken.com"),
-        "PAX":        CoinInfo(supply: 249_952_065,     startDate: "10/09/2018", website: "https://www.paxos.com/pax"),
-        "PERP":       CoinInfo(supply: 150_000_000,     startDate: nil,          website: "https://perp.fi"),
-        "PICKLE":     CoinInfo(supply: 1_289_763,       startDate: nil,          website: "https://pickle.finance"),
-        "PLOT":       CoinInfo(supply: 200_000_000,     startDate: nil,          website: "https://plotx.io"),
-        "PNK":        CoinInfo(supply: 764_626_704,     startDate: "15/03/2018", website: "https://kleros.io"),
-        "POA":        CoinInfo(supply: nil,             startDate: nil,          website: "https://poa.net"),
-        "POLS":       CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://www.polkastarter.com"),
-        "POLY":       CoinInfo(supply: 1_000_000_000,   startDate: "25/12/2017", website: "https://www.polymath.network"),
-        "PPT":        CoinInfo(supply: 53_252_246,      startDate: "12/04/2017", website: "https://populous.world"),
-        "QNT":        CoinInfo(supply: 14_612_493,      startDate: "25/06/2018", website: "https://www.quant.network/"),
-        "QSP":        CoinInfo(supply: 976_442_388,     startDate: "17/11/2017", website: "https://quantstamp.com/"),
-        "RARI":       CoinInfo(supply: 25_000_000,      startDate: "15/07/2020", website: "https://rarible.com/"),
-        "RDN":        CoinInfo(supply: 100_000_000,     startDate: nil,          website: "https://raiden.network"),
-        "REN":        CoinInfo(supply: 999_999_633,     startDate: "31/12/2017", website: "https://renproject.io"),
-        "RENBCH":     CoinInfo(supply: nil,             startDate: nil,          website: "https://renproject.io"),
-        "RENBTC":     CoinInfo(supply: nil,             startDate: nil,          website: "https://renproject.io"),
-        "RENZEC":     CoinInfo(supply: nil,             startDate: nil,          website: "https://renproject.io"),
-        "REP":        CoinInfo(supply: nil,             startDate: "17/11/2014", website: "https://www.augur.net"),
-        "REPv2":      CoinInfo(supply: nil,             startDate: nil,          website: "https://www.augur.net"),
-        "REV":        CoinInfo(supply: 1_000_000_000,   startDate: "04/08/2017", website: "https://revain.org"),
-        "RFI":        CoinInfo(supply: 10_000_000,      startDate: nil,          website: "https://reflect.finance"),
-        "RFuel":      CoinInfo(supply: nil,             startDate: nil,          website: "https://riodefi.com"),
-        "RGT":        CoinInfo(supply: nil,             startDate: nil,          website: "http://rari.capital"),
-        "RING":       CoinInfo(supply: 1_590_567_953,   startDate: "13/09/2018", website: "https://darwinia.network"),
-        "RLC":        CoinInfo(supply: 86_999_785,      startDate: "10/01/2017", website: "https://iex.ec/"),
-        "ROOT":       CoinInfo(supply: 10_000,          startDate: nil,          website: "https://rootkit.finance"),
-        "RPL":        CoinInfo(supply: 17_922_515,      startDate: "07/09/2017", website: "https://www.rocketpool.net"),
-        "RSR":        CoinInfo(supply: 100_000_000_000, startDate: nil,          website: "https://reserve.org"),
-        "RUNE":       CoinInfo(supply: 500_000_000,     startDate: nil,          website: "https://thorchain.org"),
-        "SAI":        CoinInfo(supply: nil,             startDate: "18/12/2017", website: "https://makerdao.com/en"),
-        "SALT":       CoinInfo(supply: 120_000_000 ,    startDate: "28/07/2017", website: "https://saltlending.com"),
-        "SAN":        CoinInfo(supply: 83_337_000 ,     startDate: "09/02/2017", website: "https://santiment.net"),
-        "SAND":       CoinInfo(supply: 3_000_000_000,   startDate: nil,          website: "http://www.sandbox.game"),
-        "SFI":        CoinInfo(supply: 83_284,          startDate: nil,          website: "http://saffron.finance"),
-        "SHARE":      CoinInfo(supply: 21_000_000,      startDate: nil,          website: "https://dollarprotocol.com"),
-        "SHR":        CoinInfo(supply: 4_396_000_000,   startDate: nil,          website: "https://sharering.network"),
-        "SNT":        CoinInfo(supply: 6_804_870_174,   startDate: "20/06/2017", website: "https://status.im"),
-        "SNX":        CoinInfo(supply: 190_075_446,     startDate: "07/01/2018", website: "https://www.synthetix.io"),
-        "SOCKS":      CoinInfo(supply: 326 ,            startDate: "08/05/2019", website: "https://unisocks.exchange"),
-        "SPANK":      CoinInfo(supply: 1_000_000_000 ,  startDate: "31/10/2017", website: "https://spankchain.com"),
-        "SPDR":       CoinInfo(supply: nil,             startDate: nil,          website: "https://spiderdao.io"),
-        "SRM":        CoinInfo(supply: 161_000_001,     startDate: "31/07/2020", website: "https://projectserum.com"),
-        "STAKE":      CoinInfo(supply: 8_288_374,       startDate: "31/07/2020", website: "https://xdaichain.com"),
-        "STORJ":      CoinInfo(supply: 424_999_998,     startDate: "19/07/2017", website: "https://storj.io"),
-        "SURF":       CoinInfo(supply: nil,             startDate: nil,          website: "https://surf.finance"),
-        "SUSD":       CoinInfo(supply: 22_510_653,      startDate: nil,          website: "https://www.synthetix.io"),
-        "SWFL":       CoinInfo(supply: nil,             startDate: nil,          website: "https://swapfol.io"),
-        "SWISS":      CoinInfo(supply: 10_000,          startDate: nil,          website: "https://swissfarming.finance"),
-        "SWRV":       CoinInfo(supply: 33_000_000,      startDate: nil,          website: "https://swerve.fi"),
-        "SXP":        CoinInfo(supply: 289_612_084,     startDate: "16/08/2019", website: "https://swipe.io"),
-        "Swap":       CoinInfo(supply: 100_000_000,     startDate: "09/07/2020", website: "https://trustswap.org"),
-        "TAUD":       CoinInfo(supply: nil,             startDate: nil,          website: "https://www.trusttoken.com/trueaud"),
-        "TAUDB":      CoinInfo(supply: 90_000_000_000,  startDate: nil,          website: "https://trusttoken.com"),
-        "TBTC":       CoinInfo(supply: nil,             startDate: nil,          website: "https://tbtc.network"),
-        "TCAD":       CoinInfo(supply: nil,             startDate: nil,          website: "https://www.trusttoken.com/truecad"),
-        "TGBP":       CoinInfo(supply: nil,             startDate: nil,          website: "https://www.trusttoken.com/truegbp"),
-        "THKD":       CoinInfo(supply: nil,             startDate: nil,          website: "https://www.trusttoken.com/truehkd"),
-        "THKDB":      CoinInfo(supply: 90_000_000_000,  startDate: nil,          website: "https://trusttoken.com"),
-        "TKN":        CoinInfo(supply: 39_406_760,      startDate: "25/04/2017", website: "https://monolith.xyz"),
-        "TRB":        CoinInfo(supply: 1_622_063,       startDate: "10/09/2019", website: "https://tellor.io"),
-        "TRST":       CoinInfo(supply: 100_000_000,     startDate: "07/12/2016", website: "https://www.wetrust.io"),
-        "TRU":        CoinInfo(supply: 1_446_312_655,   startDate: nil,          website: "https://truefi.io"),
-        "TUSD":       CoinInfo(supply: nil,             startDate: "31/03/2018", website: "https://www.trusttoken.com"),
-        "TUSDB":      CoinInfo(supply: 90_000_000_000,  startDate: nil,          website: "https://trusttoken.com"),
-        "UBT":        CoinInfo(supply: 150_000_000,     startDate: "11/12/2017", website: "https://unibright.io"),
-        "UMA":        CoinInfo(supply: 101_104_951,     startDate: "09/01/2020", website: "https://umaproject.org"),
-        "UNI":        CoinInfo(supply: 1_000_000_000,   startDate: nil,          website: "https://uniswap.org"),
-        "USDC":       CoinInfo(supply: nil,             startDate: "17/05/2018", website: "https://www.centre.io/usdc"),
-        "USDN":       CoinInfo(supply: 125_679_184,     startDate: nil,          website: "https://neutrino.at"),
-        "USDS":       CoinInfo(supply: 90_000_486_303,  startDate: "15/08/2018", website: "https://www.stably.io"),
-        "USDT":       CoinInfo(supply: 9_479_177_442,   startDate: "10/06/2014", website: "https://tether.to"),
-        "USDx":       CoinInfo(supply: nil,             startDate: nil,          website: nil),
-        "UTK":        CoinInfo(supply: 500_000_000,     startDate: nil,          website: "https://utrust.com"),
-        "VERI":       CoinInfo(supply: 100_000_000,     startDate: "25/04/2017", website: "https://veritas.veritaseum.com"),
-        "WAVES":      CoinInfo(supply: nil,             startDate: nil,          website: "https://waves.tech"),
-        "WBTC":       CoinInfo(supply: 21_000_000,      startDate: "27/11/2018", website: "https://wbtc.network"),
-        "WETH":       CoinInfo(supply: 1_153_917,       startDate: "31/12/2018", website: "https://weth.io"),
-        "WFIL":       CoinInfo(supply: nil,             startDate: nil,          website: "https://www.wrapped.com"),
-        "WHITE":      CoinInfo(supply: nil,             startDate: nil,          website: "https://www.whiteheart.finance"),
-        "WICC":       CoinInfo(supply: 210_000_000,     startDate: nil,          website: "https://www.waykichain.com"),
-        "WISE":       CoinInfo(supply: nil,             startDate: nil,          website: "https://wisetoken.net"),
-        "WRX":        CoinInfo(supply: 995_833_334,     startDate: nil,          website: "https://wazirx.com"),
-        "WTC":        CoinInfo(supply: 70_000_000,      startDate: "27/08/2017", website: "https://www.waltonchain.org/en"),
-        "WZEC":       CoinInfo(supply: 9_082,           startDate: nil,          website: "https://www.wrapped.com"),
-        "XCHF":       CoinInfo(supply: 100_000_000,     startDate: "01/11/2018", website: "https://swisscryptotokens.ch"),
-        "XFT":        CoinInfo(supply: 10_000_000,      startDate: nil,          website: "https://offshift.io"),
-        "XIO":        CoinInfo(supply: 100_000_000,     startDate: "22/10/2019", website: "https://xio.network"),
-        "XOR":        CoinInfo(supply: 350_000,         startDate: nil,          website: "https://sora.org"),
-        "XRP":        CoinInfo(supply: 99_990_932_726,  startDate: "02/02/2013", website: "https://ripple.com/xrp"),
-        "XRT":        CoinInfo(supply: 9_984_038,       startDate: nil,          website: "https://robonomics.network/en/"),
-        "YAX":        CoinInfo(supply: 305_540,         startDate: nil,          website: "https://yaxis.io"),
-        "YFI":        CoinInfo(supply: 30_000,          startDate: nil,          website: "https://yearn.finance"),
-        "YFII":       CoinInfo(supply: 39_375,          startDate: nil,          website: "https://dfi.money"),
-        "YFIM":       CoinInfo(supply: 60_000,          startDate: nil,          website: nil),
-        "Yf-DAI":     CoinInfo(supply: 21_000_000,      startDate: nil,          website: "https://www.yfdai.finance"),
-        "ZAI":        CoinInfo(supply: nil,             startDate: nil,          website: "https://zai.finance"),
-        "ZEC":        CoinInfo(supply: 21_000_000,      startDate: "28/10/2016", website: "https://z.cash"),
-        "ZRX":        CoinInfo(supply: 1_000_000_000,   startDate: "15/08/2017", website: "https://www.0xproject.com/#home"),
-        "eXRD":       CoinInfo(supply: 4_410_000_000,   startDate: nil,          website: "https://www.radixdlt.com"),
-        "imBTC":      CoinInfo(supply: 21_000_000,      startDate: nil,          website: "https://tokenlon.im/imbtc#"),
-        "pBTC":       CoinInfo(supply: 21_000_000,      startDate: nil,          website: "https://ptokens.io"),
-        "sETH":       CoinInfo(supply: 29_935,          startDate: nil,          website: "https://www.synthetix.io"),
-        "sXAU":       CoinInfo(supply: 225.36884553,    startDate: nil,          website: "https://www.synthetix.io"),
-        "wANATHA":    CoinInfo(supply: nil,             startDate: nil,          website: "https://anatha.io"),
-
-        "FTT":        CoinInfo(supply: 345_219_294,     startDate: nil,          website: "https://ftx.com"),
-        "MX":         CoinInfo(supply: 608_002_217,     startDate: nil,          website: "https://www.mxc.ceo"),
-        "GT":         CoinInfo(supply: 300_000_000,     startDate: nil,          website: "https://gatechain.io"),
-
-    ]
-
-}
-
-struct CoinInfo {
-    let supply: Decimal?
-    let startDate: String?
-    let website: String?
-}
