+++ UnstoppableWallet/UnstoppableWallet/Modules/Coin/CoinChart/Views/ChartCurrentRateCell.swift
