+++ UnstoppableWallet/UnstoppableWallet/Modules/Coin/CoinChart/Views/ChartIntervalAndSelectedRateCell.swift
