+++ UnstoppableWallet/UnstoppableWallet/Modules/Coin/CoinChart/Views/MacdPointView.swift
