+++ UnstoppableWallet/UnstoppableWallet/Modules/Coin/CoinChart/Views/MarketInfoCell.swift
