+++ UnstoppableWallet/UnstoppableWallet/Modules/Coin/CoinChart/Views/PriceIndicatorCell.swift
