+++ UnstoppableWallet/UnstoppableWallet/UserInterface/CurrencyCompactFormatter.swift
